requirements.txt
Django==5.2.6

django-filter==25.2

djangorestframework==3.16.1

djangorestframework_simplejwt==5.5.1